feat: add Base62 URL Shortener strategy
