+++ lib/dynamic_links/shortening_strategies/base_strategy.rb
@@ -1,0 +1,10 @@
+module DynamicLinks
+  module ShorteningStrategies
+    class BaseStrategy
+      def shorten(url)
+        raise NotImplementedError, "You must implement the shorten method"
+      end
+    end
+  end
+end
+
