+++ lib/dynamic_links/shortening_strategies/md5_strategy.rb
@@ -1,0 +1,9 @@
+module DynamicLinks
+  module ShorteningStrategies
+    class Md5Strategy < BaseStrategy
+      def shorten(url)
+        # Implementation for MD5 shortening
+      end
+    end
+  end
+end
